Reorganise imports for project-builder
