+++ project/config.py
@@ -1,0 +1,8 @@
+import os.path
+
+ROOTDIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
+
+TEMPLATEDIR = os.path.join(ROOTDIR, 'templates')
+
+# Blacklist of projects to ignore.
+IGNORED = []
